+++ config.py
@@ -20,8 +20,8 @@ import threading
 import uuid as _uuid
 from pathlib import Path
 
-VERSION = "2.30.8"
-BUILD_DATE = "2026-03-13"
+VERSION = "2.31.0"
+BUILD_DATE = "2026-03-14"
 
 __all__ = [
     "BUILD_DATE",